fix: move telemetry modules to agents/ root, fix architecture boundaries
Move skill-bins.ts and usage-tracker.ts from agents/skills/ to agents/
to resolve import boundary violations. bash-tools.exec.ts can now safely
import from sibling modules.
