+++ src/cli/skills-cli.ts
@@ -20,8 +20,8 @@ import { normalizeOptionalString } from "../shared/string-coerce.js";
 import { formatDocsLink } from "../terminal/links.js";
 import { theme } from "../terminal/theme.js";
 import { renderTable } from "../terminal/table.js";
-import { getSkillUsage } from "../agents/skills/usage-tracker.js";
-import type { SkillUsageEntry } from "../agents/skills/usage-tracker.js";
+import { getSkillUsage } from "../agents/usage-tracker.js";
+import type { SkillUsageEntry } from "../agents/usage-tracker.js";
 import { CONFIG_DIR } from "../utils.js";
 import { resolveOptionFromCommand } from "./cli-utils.js";
 import { formatSkillInfo, formatSkillsCheck, formatSkillsList } from "./skills-cli.format.js";
